moving RAC_CMD and cron to lib
